Bug 22664: Link names in order search results instead of numbers
To test:
Make up some orders, put some in basket groups.
Search for your vendor's orders, note that the only thing linking to the
order is a tiny number in parentheses.
Apply this patch, restart the things.
Refresh the results, note that te basket names *and* numbers are linked
now.
Rejoice.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Séverine QUEUNE <severine.queune@bulac.fr>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>