Bug 22537: Suspend all hold button appears even if the hold can no longer be suspended.
This patch adds to the user summary page in the OPAC a check for waiting
or in-transit holds. If all holds are waiting or in transit the "Suspend
all" controls will be hidden.

This patch also adds a missing $KohaDates filter to the "suspended
until" date.

To test, apply the patch and log in to the OPAC as a patron with holds.
Test these conditions:

 - All holds are waiting and/or in transit
   - The "Suspend holds" controls at the bottom of the page should not
     appear.
 - Some holds are waiting and/or in transit
   - The "Suspend holds" controls should appear.

Signed-off-by: Marjorie <marjorie.barry-vila@collecto.ca>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>