Bug 22803: (alternate) Set dataTable width issue
This patch sets the "autowidth" option to "false" for the OPAC
subscriptions DataTable on opac-detail.pl. This gives the table a fluid
width rather than having the width be "hard-coded" by DataTables upon
first page render.

To test, apply the patch and view the detail page for a bibliographic
record with a subscription. Under the "Subscriptions" tab, the table of
latest issues should fill the width of the tab container.

Resize the browser window and confirm that the table width adjusts
accordingly.

Signed-off-by: Liz Rea <wizzyrea@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>