Bug 22877: (QA follow-up) Clarify intent of tests and remove one.
This patch embelishes the existing comments slightly and removes a test
which had a FIXME, but was in fact just a duplicate of the test directly
above it.  I beleive it became a duplicate after the subroutine signature
changed at some point recently.  I also move the bug 22877 specific test
into a subtest.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Liz Rea <wizzyrea@gmail.com>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>