Bug 22884: Remove field ending . from XSLT templates
Currently Koha very inconsistently adds a . at the end when
outputting MARC fields on the detail views.

This patch removes the few cases where we add in .
Keeping the end user in mind, I believe it's not needed and
the added consistency will be nice. ISBD should be added on
cataloguing or by templates, but we should not mix when possible.

To test:
Catalog the following fields, if repeatable, add more than one:
- 020
- 022
- 111
- 242
- 245
- 246
- 250
- 300
- 440
- 490
- 700/710/711 with $i
- 700/710/711 ind.2 = 2
Verify both on staff and OPAC detail pages that there is no
additional . at the end of the line.

Signed-off-by: Hayley Mapley <hayleymapley@catalyst.net.nz>
Signed-off-by: frederik <frederik@inlibro.com>
Signed-off-by: Hayley Mapley <hayleymapley@catalyst.net.nz>
Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>