Bug 22901: On items search fix search field change listening
In items search, when there is a custom field with authorised values :
when changing the conditional select AND/OR then the authorised values select disappears.
This behaviour is like when the search field is changed.

The bug is that JavaScript code meaning search field change listens on all selects instead of only search field select.

Test plan :
1) Define a custom items search field with authorised values
2) Go to items search
3) Click on 'New field'
4) On second line, select custom field
5) The authorised values select appears
6) Select 'AND' in conditional select
7) Without patch the authorised values select disappears, with patch nothing changes
8) On second line, select a non-custom field
9) The authorised values select is replaced by a text field
10) On second line, select custom field
11) The authorised values select appears

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>