Bug 22954: Minor markup error in OPAC messaging template
This patch corrects a minor error in opac-messaging.tt in order to
quiet errors returned by the HTML validator.

To reproduce the problem you should have the
EnhancedMessagingPreferencesOPAC  system preference enabled. Log in
to the OPAC and go to the "your messaging" page. Run the page through
the W3C validation service: https://validator.w3.org. Either by pasting
in the URL for the search results (if web accessible) or by viewing
source, copying, and pasting into "Validate by Direct Input."

This patch addresses this error:

 - "Self-closing syntax (/>) used on a non-void HTML element. Ignoring
   the slash and treating as a start tag."

To test, apply the patch and try again to reproduce the validation
error. The error should no longer be present. There should be no
visible change to the page.

Signed-off-by: Nadine Pierre <nadine.pierre@inLibro.com>
Signed-off-by: Arthur Bousquet <arthur.bousquet@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>