Bug 22955: Markup error in OPAC lists template
This patch makes a minor correction to the template for OPAC lists.
This correction fixes an error returned by the HTML validator by adding
an empty "alt" attribute to an image tag.

To reproduce the problem you should have at least one list available in
the OPAC with one or more items on it. View the contents of that list in
the OPAC. Run the page through the W3C validation service:
https://validator.w3.org. Either by pasting in the URL for the search
results (if web accessible) or by viewing source, copying, and pasting
into "Validate by Direct Input."

This patch addresses one specific error:

 - "An img element must have an alt attribute, except
    under certain conditions."

To test, apply the patch and try again to reproduce the validation
error. The error should no longer be present. There should be no
visible change to the page.

Click "Download list." A modal window should appear with a "Loading"
message, followed by a form for selecting a download format.

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>