Bug 23061: Add column/print/export feature on checkout history page
This patch adds columns configuration and export features to the
checkout history page.

This patch also removes some inline CSS: "overflow:hidden" from both the
checkout and holds history pages. This was breaking the table controls
tooltips, and didn't appear to me to be necessary.

To test, apply the patch and restart Plack (or restart_all on
kohadevbox).

 - Open a patron's checkout history page in the staff client.
 - The table of checkouts should have controls at the top.
   for columns configuration and export/print.
 - These controls should work correctly.

Signed-off-by: Michal Denar <black23@gmail.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>