Bug 8260: user rank order instead of absolute rank to set dropdowns on routing.pl
When deleting a borrower on a routing list we leave a gap in the numbering. It appears to only be a problem when viewing
or editing the list in the staff client.

Rather than use the rank in the DB to set the values, we use the implicit order

To test:
 1 - Create a subscription (or find one)
 2 - Add a routing list
 3 - Add five patrons
 4 - Delete the patron who is third in the list
 5 - Edit the routing list
 6 - Note the priority for the last patron is 1
 7 - Apply patch
 8 - Reload the page
 9 - Values look correct
10 - Confirm reordering works as expected

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>