Bug 23279: In news management interface, sort news by publication date descending
This patch makes a few changes to the configuration of the DataTable
showing news items in the staff client. It adds a default sort on
"Publication date," descending; It updates the way non-sortable columns
are configured; It modifies sorting on the title and content columns so
that articles are ignored.

To test, apply the patch and go to Tools -> News.

 - When the page loads the table of news items should sort by default on
   publication date, descending.
 - There should be no sorting indicator on the first and last columns.
 - When sorting the "Title" or "News" columns, articles "A," "An," and
   "The" should be ignored.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>
Signed-off-by: frederik chenier <frederik.chenier@inlibro.com>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>