Bug 22709: (RM follow-up) AddBiblio returns a list not a scalar
AddBiblio was being called in scalar context in
Biblio_and_Items_plugin_hooks.t and this was leading to difficult to
reproduce test failures. If the tests was run after any other test that
also called AddBiblio then it would fail, but if it was run as the first
test it would pass. This is due to a fluke in processing which meant the
the final return value of the list (itemnumber) happened to match a
biblionumber.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>